Handling of conflicts and dependencies at the backend
OCI can now resolve conflicts and package dependencies at the backend prior to the installation of packages. The process is pretty straightforward -

backend.cpp -> Register the service org.opensuse.OCIHelper and create a proxy for org.opensuse.oneclickinstaller dbus service
Step 1. Mark packages for installation in the pool
Step 2. Resolve conflicts. Repeat this step until there are no more conflicts.
Step 3. Commit changes to the system (this is where the actual installation occurs)
The communication between the OCI and OCIHelper is asynchronous [and semi-async. check line 93].
=====
Updated the .xml files as per the requirement (they are prone to change forever :))
conflictresolutionscreen.h.cpp - modified them to handle the signals from OCIHelper and show GUI accordingly.
=====
After the noConflicts() signal is emitted from OCIHelper to OCI, please notice the change in the header.
Should change utils.h/.cpp to handle all the packages and repos mentioned in the .ymp file (that will be my next commit)
But for now, backend.cpp is hard-coded to handle vlc package from the vlc-repo. This is just for a test :)